star/LGESQL/sparc/requirements.txt
dgl-cu111==0.6.1
embeddings==0.0.8
transformers==4.18.0
stanza==1.1.1
nltk==3.5
sqlparse==0.4.2